NetStats sent/received improvements
Shows digits now, to be fixed: from 99.5 it rounds up to 100 and
displays that

+++ StatProgramProject/StatProgramProject/Uptime.cs
@@ -10,7 +10,7 @@ namespace StatProgramProject
 {
     public partial class MainForm:Form
     {
-        private const double UPTIME_TIMER_UPDATE = 1000;
+        private const int UPTIME_TIMER_UPDATE = 1000;
         private Timer uptimeTimer;
         long totalSeconds = 0;
         int weeks = 0, days = 0, hours = 0, minutes = 0, seconds = 0;
